feat: new table and user layer

+++ src/tasks/tasks.service.ts
@@ -1,34 +1,39 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { Task } from './entities/task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { PrismaService } from '../prisma/prisma.service'
 import { PaginationDto } from 'src/common/dto/pagination.dto';
 
 @Injectable()
 export class TasksService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private prisma: PrismaService) { }
 
   async findAll(paginationDto?: PaginationDto) {
-    const { limit = 10, offset = 0 } = paginationDto ?? {};
+    const { limit = 10, offset = 0 } = paginationDto;
 
     const allTasks = await this.prisma.task.findMany({
       take: limit,
       skip: offset,
+      orderBy: {
+        createdAt: "desc"
+      }
     });
+
     return allTasks;
   }
 
-  async findOne(id: string) {
+  async findOne(id: number) {
     const task = await this.prisma.task.findFirst({
       where: {
-        id: Number(id),
-      },
-    });
+        id: id
+      }
+    })
 
     if (task?.name) return task;
 
-    throw new HttpException('Tarefa não encontrada...', HttpStatus.NOT_FOUND);
+    throw new HttpException("Tarefa não foi encontrada!", HttpStatus.NOT_FOUND)
+
   }
 
   async create(createTaskDto: CreateTaskDto) {
@@ -38,69 +43,73 @@ export class TasksService {
           name: createTaskDto.name,
           description: createTaskDto.description,
           completed: false,
-        },
-      });
+          userId: createTaskDto.userId
+        }
+      })
+
       return newTask;
-    } catch (erro) {
-      throw new HttpException(
-        'Falha ao criar uma tarefa',
-        HttpStatus.BAD_REQUEST,
-      );
+    } catch (err) {
+      console.log(err);
+      throw new HttpException("Falhaa ao cadastrar trarefa", HttpStatus.BAD_REQUEST)
     }
   }
 
-  async update(id: string, updateTaskDto: UpdateTaskDto) {
+  async update(id: number, updateTaskDto: UpdateTaskDto) {
     try {
       const findTask = await this.prisma.task.findFirst({
         where: {
-          id: Number(id),
-        },
-      });
+          id: id
+        }
+      })
 
       if (!findTask) {
-        throw new HttpException(
-          'Essa tarefa não existe!',
-          HttpStatus.NOT_FOUND,
-        );
+        throw new HttpException("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
       }
+
 
       const task = await this.prisma.task.update({
         where: {
-          id: findTask.id,
+          id: findTask.id
         },
-        data: updateTaskDto,
-      });
+        data: {
+          name: updateTaskDto?.name ? updateTaskDto?.name : findTask.name,
+          description: updateTaskDto?.description ? updateTaskDto?.description : findTask.description,
+          completed: updateTaskDto?.completed ? updateTaskDto?.completed : findTask.completed,
+        }
+      })
 
       return task;
-    } catch (erro) {
-      console.log('erro: ' + erro);
-      throw new HttpException(
-        'Falha ao editar a tarefa',
-        HttpStatus.BAD_REQUEST,
-      );
+    } catch (err) {
+      throw new HttpException("Falha ao atualizar essa tarefa", HttpStatus.BAD_REQUEST)
     }
   }
 
-  async delete(id: string) {
+
+  async delete(id: number) {
     try {
-      const findTask = this.prisma.task.findFirst({
+      const findTask = await this.prisma.task.findFirst({
         where: {
-          id: Number(id),
-        },
-      });
+          id: id
+        }
+      })
+
+      if (!findTask) {
+        throw new HttpException("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
+      }
+
       await this.prisma.task.delete({
         where: {
-          id: (await findTask).id,
-        },
-      });
+          id: findTask.id
+        }
+      })
+
       return {
-        message: 'Tarefa deletada com sucesso!',
-      };
-    } catch (erro) {
-      throw new HttpException(
-        'Falha ao deletar a tarefa',
-        HttpStatus.BAD_REQUEST,
-      );
+        message: "Tarefa deletada com sucesso!"
+      }
+    } catch (err) {
+      throw new HttpException("Falha ao deletar essa tarefa", HttpStatus.BAD_REQUEST)
     }
+
   }
+
 }
